+++ effect_wave.ino
@@ -9,7 +9,7 @@ void waveEffect(Adafruit_NeoPixel& strip, uint32_t color) {
     if (millis() - lastUpdate >= interval) {
         lastUpdate = millis();
 
-        for (int i = 0; i < strip.numPixels(); i++) {
+        for (uint16_t i = 0; i < strip.numPixels(); i++) {
             float brightness = (sin((i + waveOffset) * 0.2) + 1.0) * 127.5; // Sine wave from 0 to 255
             strip.setPixelColor(i, strip.Color((brightness * ((color >> 16) & 0xFF)) / 255,
                                                (brightness * ((color >> 8) & 0xFF)) / 255,
